Trying to fix tests
